JsonモデルをVueValueから分離、model.jsonへ配置
